move shared middleware out of common.ts
